convert doesn't initialize CLIPythonManager, other bit more factoring
