+++ vcell-cli/src/main/java/org/vcell/cli/run/ExecuteCommand.java
@@ -159,7 +159,7 @@ public class ExecuteCommand implements Callable<Integer> {
             for (String simName : resultsHash.keySet()) {
                 String CSVFilePath = Paths.get(outDirForCurrentVcml.toString(), simName + ".csv").toString();
                 RunUtils.createCSVFromODEResultSet(resultsHash.get(simName), new File(CSVFilePath));
-                RunUtils.transposeVcmlCsv(CSVFilePath);
+                PythonCalls.transposeVcmlCsv(CSVFilePath);
             }
         } catch (IOException e) {
             System.err.println("IOException while processing VCML " + vcmlFile.getName());
@@ -227,7 +227,7 @@ public class ExecuteCommand implements Callable<Integer> {
         }
 
         RunUtils.removeAndMakeDirs(new File(outputDir));
-        RunUtils.generateStatusYaml(inputFileName, outputDir);    // generate Status YAML
+        PythonCalls.generateStatusYaml(inputFileName, outputDir);    // generate Status YAML
 
         // from here on, we need to collect errors, since some subtasks may succeed while other do not
         // we now have the log file created, so that we also have a place to put them
@@ -401,7 +401,7 @@ public class ExecuteCommand implements Callable<Integer> {
                     logDocumentMessage += "Done. ";
                 }
 
-                genPlotsPseudoSedml(sedmlLocation, outDirForCurrentSedml.toString());    // generate the plots
+                PythonCalls.genPlotsPseudoSedml(sedmlLocation, outDirForCurrentSedml.toString());    // generate the plots
                 oneSedmlDocumentSucceeded = true;
             } catch (Exception e) {
                 somethingFailed = true;
@@ -500,20 +500,5 @@ public class ExecuteCommand implements Callable<Integer> {
                     StandardOpenOption.CREATE, StandardOpenOption.APPEND);
         }
     }
-
-    // Due to what appears to be a leaky python function call, this method will continue using execShellCommand until the unerlying python is fixed
-    private static void genPlotsPseudoSedml(String sedmlPath, String resultOutDir) throws PythonStreamException, InterruptedException, IOException {
-        CLIPythonManager.callNonsharedPython("genPlotsPseudoSedml", sedmlPath, resultOutDir);
-//        ProcessBuilder pb = new ProcessBuilder(new String[]{CLIResourceManager.python, cliDirs.cliPath.toString(), "genPlotsPseudoSedml", sedmlPath, resultOutDir});
-//        runAndPrintProcessStreams(pb, "","");
-
-        /**
-         * replace with the following once the leak is fixed
-         */
-//        CLIPythonManager cliPythonManager = CLIPythonManager.getInstance();
-//        String results = cliPythonManager.callPython("genPlotsPseudoSedml", sedmlPath, resultOutDir);
-//        cliPythonManager.printPythonErrors(results);
-    }
-
 
 }
